requirements.txt
Flask
Flask-Login
psycopg2-binary
werkzeug
python-dotenv
requests